move range serialization tests to kernel tests
